implemented basedatastore for franchises
