+++ GTBA/GTBA/Services/IFranchisesDataStore.cs
@@ -1,0 +1,12 @@
+﻿using GTBA.Models;
+using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace GTBA.Services
+{
+    public interface IFranchisesDataStore : IDataStore<Franchise>
+    {
+
+    }
+}
